TANI/UstSolunumYolu/modules/vision_cxr_covid/requirements.txt
tensorflow-macos>=2.15.0
tensorflow-metal>=1.1.0
kagglehub>=0.2.0
pillow>=10.0.0
pyyaml>=6.0.1
numpy>=1.24.0
